MIPS: Resolve "relocation overflow" when using gold linker
The "relocation overflow" error appears because of branching to
global labels. This error doesn't appear when using ld.bfd because
that linker doesn't handle this case the same way as gold does.
Gold linker creates a PLT which is far away from the calling
function, which results in a "relocation overflow". To avoid
creating the PLT, branch to a non-global label instead.

Change-Id: Iff54092bb0c2455a4eaff68a5cd536b3f2b819b3
Signed-off-by: Dejan Jovicevic <dejan.jovicevic@mips.com>