MIPS: Resolve "relocation overflow" when using gold linker
The "relocation overflow" error appears because of branching to
global labels. This error doesn't appear when using ld.bfd because
that linker doesn't handle this case the same way as gold does.
Gold linker creates a PLT which is far away from the calling
function, which results in a "relocation overflow". To avoid
creating the PLT, branch to a non-global label instead.

Built and booted mips32r2 and mips64r6 emulators, art tests
are passing.

Change-Id: I95bf11af95f788ede585c313a305f755cb2add29
Signed-off-by: Dejan Jovicevic <dejan.jovicevic@mips.com>