MIPS32: Fix invalid $gp value in READ_BARRIER
After multiple consecutive READ_BARRIER calls, sometimes
the $gp register ends up holding an invalid value. Avoid
this by saving the $gp value on the stack at the beginning
and restoring it at the end of READ_BARRIER macro.

This issue appeared when Android was built with the gold linker.
Tests that were failing: 071, 107 and 151

Change-Id: Iacbbe9a90a29c6eb9b26ba773f2f028de50a8f48
Signed-off-by: Dejan Jovicevic <dejan.jovicevic@mips.com>