MIPS32: Fix invalid $gp value in READ_BARRIER
After multiple consecutive READ_BARRIER calls, sometimes
the $gp register ends up holding an invalid value. Avoid
this by saving the $gp value on the stack at the beginning
and restoring it at the end of READ_BARRIER macro.

This issue appeared when Android was built with the gold linker.
Tests that were failing: 071, 107 and 151

Change-Id: I34f6ef416b36bdad094ba202c6ebd608a062eb15
Signed-off-by: Dejan Jovicevic <dejan.jovicevic@mips.com>